Evaluate the v_over_s face value properly (rather than a FE interpolation of everything) for the CFL number, although it looks very similar and didnt help the Ultimate limiter. Perhaps Ultimate isnt suitable for this application so I wouldnt use it.
Correctly use a 'ggi' to indicate the quadrature point value of gradient 
of pressure, rather than '1'. Doesnt actually matter with linear pressure 
as the gradient is element wise, but to go higher it would.

For saturation subcycling actually find the upwind values for the latest 
old_sfield_subcycle values rather the old_sfield values.

The test case .diml has a trivial change.