Minor change to femtools/CV_Face_Values.F90 to make more procedures public.
Add a script to compare the numerical answer to the analytic on the fine 
analytic mesh.

Adjust code in many ways I cant remember!! LP diff will remind me ...

Probably to do with consistent subcycling, slight adjustment of options, 
permit it with a diagnostic field. Fix a couple of minor errors 
associated with the initialisation.