Change the relperm fields to be defined by an internal correlation rather than python diagnostic such as to be able to find the values at quadrature points and boundarys. Adjust code and tests as required.
Make the options input for gravity optional. If not there allocated 
a constant field that is zeroed.

Add order of convergence and checks and regression checks for the corey 
test case. Just noticed that gmsh isnt producing same mesh every time 
so checking against reference solutions in vtus is complaining. Part 
way to fixing this by commiting the gmsh output .msh files. Just need 
to regenerate the regression solutions.