Adjust face value schemes for relative permeability to add effectively the v_over_s method into this branch.
This addes capability to pick the relperm face value from the 
FE interpolation and the correlation. Both have shown to not 
converge to the correct answer and will be removed in a following 
commit (just want to have it backed in revisio control on LP).

Add more checks to the single phase case.