Remove test case that perhaps isnt physically correct. To be replaced via changing the working test case in small stages.
Change the PowerLaw relperm correlation to require a real list/vector 
of values such that each phase can have a different exponent.

Adjust the single phase test case as required as it uses this (albeit 
with the exponents zeroed, but it atleast checks it reads the list 
and uses them).