Hide the GradientPressure field and meshes. This only exists as for now there is no function like face_grad_at_quad (note there is a ele_grad_at_quad).
Permit options to choose the pressure mesh order and the quadrature order.
This so one can do quadratic basis functions (well for what works in 
Fluidity for P2 CV, which isnt as robust as P1 CV so dont use it yet).

Change the saturation source to be on the element wise mesh as it 
is useful to permit discontinuities for this.

Allow the viscosity field to be python diagnostic as well as prescribed.

Add another pseudo 1 phase test case to check velocity BC at the inlet.

Add a 2 phase quadratic relperm with gravity 'updip'. Requires analytic solution, 
options to match this and regresssion solutions generating.