Remove pressure BC options from phases >= 2, as they are not required.
Add an option to specify the number of non linear iterations for 
the first time step. This is proven useful for the three phase in 
layers case where 2 and 3d require 4 iterations for the first timestep 
then 1 after.

Adjust the inputs for the three phase density driven flow. xml file
requires setting up to actually test something.