Make sure the timestep does not overshoot the finish time. This was causing issues with adaptive time step to a desired finish time to compare to a analytic reference.
This change, along with a previous that defaulted the max dt change 
permitted to 10% increase, has implied that all the regression tests 
need re-running and storing. To make this easier all tests are set up 
to only output _0.vtu and _1.vtu.

Also, change the name of the DarcyVelocityCFL field to 
DarcyVelocityOverPorosityCFL as that is what the code was 
doing and after checking is a suitable number.