Move non-system linker paths to start of search path
Linker paths specified with -L are searched in order for a library.
To ensure that local paths are searched before system ones, move
BLAS_LIBS to the end of LIBS in the two remaining Makefiles that
didn't already do this.  Also, prepend the PETSc link line to LIBS
rather than appending it.  This way if there is a version of PETSc
installed in /usr/lib (as is the case for ubuntu systems), we don't
pick it up at link time even though PETSC_DIR and PETSC_ARCH have been
set appropriately.