Add a diagnostic to phase1 that calculates the bulk darcy velocity = sum (S_i * V_i)
Add options for metric advection back for darcy model, only 
permit  prescribed number of subcycling and add a new option 
to the general Fluidity routine to specify the advecting velocity 
to use. Hard set in the darcy options this advecting velocity 
name to BulkDarcyVelocity.

This is tested yet so shouldnt be used.

Update test cases for bulk velocity field.