Fix some space style in Darcy_IMPES_Assemble
In Populate_State allocate the necessary fields for dual perm/poro 
Like prime per/poro fields this inserts the fields in state(1) and 
then alias in all other states to these.

This requires a different name for the dual perm and porosity such 
that the schema has changed. I could not see anyother way around this.

This will need to be accounted for in the darcy initialise routine which 
can easily be done with a logical argument which will set a post append 
character to either "Dual" or "".