Change the pressure assemble to use the petsc_csr_matrix.
Adjust femtools to include a petsc_solve routine that will take
the petsc_csr_matrix and a scalar_field_pointer array. Similar 
adjustments for the dirichlet BC's. Include in Petsc_Tools functions 
to copy to and from a petsc vector and a scalar_field_pointer array.

Seems to work fine for 1d and 3d. For 2d although the answer may be correct
petsc keeps complaing about trying to insert a number into a row-column 
it doesnt know about. This can start to happen part way through 
into the simulation (after a few timesteps). This slows it down alot. Looks 
like a bug, just cannot see if this is in the Darcy code, femtools or petsc.