Add functionality to Sparse_Tools_Petsc.F90 to add a scalar field to the diagonal of a specified block of a petsc_csr_matrix
Add to darcy impes the lambda*cv_mass matrix in the pressure equation 
and the saturation sources for dual model. These terms assume pressure is 
control volume wise, or that a lumped mass matrix is used. A full consideration 
of pressure as FE wise requires functionality not there in femtools.