Correct-looking solution in Dirichlet BC implementation.
In order to apply a Dirichlet BC, the algorithm (not implemented in this
order in the code, but can most simply be understood as):

- Determine which elements are on the given boundary by picking those in
  the surface mesh whose value in mesh.faces.boundary_ids match the 
  given boundary id.
- Create a set of these elements and use mesh.faces.surface_node_list to
  build a mapping from surface elements to global nodes.
- Zero the rows except the main diagonal of the rows whose row number is
  in the image of this mapping
- Execute the apply_dirichlet kernel iterating over the surface elements,
  accessing the node values through the surface_elements -> global_nodes
  mapping.

This produces a correct-looking solution for the Laplace example. The next
step is to convert this test into a convergence test to ensure correctness,
and generalise the code for the application of BCs.