Don't call op2.init unless we see a UFL equation
The user may have floppy correctly installed, but not run any UFL
equations, so add an _init method in floppy that is called before
trying to run a UFL equation.  This means we can't initialise the
dirichlet BC kernel in solving at file level, so just do it when
needed.