Support for exterior facet integrals.
Exterior facet integrals are assembled in a similar manner to cell
integrals, except the iteration set is the set of cells on the boundary.
The dat of facet IDs is passed so that the FFC-generated kernel can
choose the correct DoFs for assembly.

The ufl_weak_bcs test tests exterior facet integrals with the
implementation of a Neumann boundary condition for Laplace's equation.