Fix coplanar ids calculation for internal boundaries.
This fixes https://bugs.launchpad.net/fluidity/+bug/1088860 (hopefully) which prevented make_eelist from giving 
an error when elements were adjacent to more than one elements through the same facet. Fixing this means we can
no longer call make_eelist on the surface_mesh if there is an internal boundary. Instead in the coplanar calculation
we now construct a temporary copy of the surface mesh with one of each pair of duplicate facets on the internal boundary
removed.