On second thought, don't zero columns in ctp_m/ct_m. Do enforce bcs in velocity correction solve.
Lifting the dirichlet bcs from the continuity equation by zeroing out columns 
in ctp_m and adding the contributions into ct_rhs turns out not to work. As
ct_rhs gets reassembled every nl. it. In next iteration we can't recompute
the lifted contributions to ct_rhs as the corresponding columns in ctp_m are
already zeroed. Instead we now only zero the columns of ctp_m and ct_m when
copying them into petsc matrices inside full_projection. The pressure projection
equation already has the right lifted values on the rhs because of the 
ctp_m u* term which gives the right terms as u^* already satisfies the bcs
and we haven't zeroed the corresponding columns. We do now have to apply
bcs to the velocity correction to avoid the nonzeroed rows of ct_m^T
clobbering the velocity bcs.