testharness: Only run genpbs if asked
Previously some complicated logic was followed whether to run genpbs or not (nprocs>1 or length=="long" and not --no_pbs)
With this commit genpbs is only called if --genpbs is specified. The --no_pbs option is therefore superseded.
This will allow setting the correct NPROCS for short and medium tests without breaking "make test" on our normal ubuntu machines.