Fix and simplify version info script.
The script no longer uses configure time variables, so doesn't need to be
created from tools/version-info.in anymore. To avoid conflicts, instead
of committing it as tools/version-info, I have renamed it to
tools/git-version-info. The script now only listens to the PACKAGE_VERSION
env. variable. Passing on the configure time FLUIDITY_VERSION is taken
care of in the Makefile. You can still override it by running:
   make PACKAGE_VERSION=foo
The script is no longer installed in the local bin/ and therefore also
not during "make install". This script really doesn't make sense outside
the context of the source directory. Running it directly from $srcdir/tools
has the added advantage that it's easy to workout the directory git needs
to look at to work out the version string using $(dirname $0).