Rescale diagonal pivots for lifted boundary conditions.
This is to work around an issue with GAMG on a matrix where the blocksize is set - where
the pivot value may change the strong connection criterion for the entire block - thus a strong
free slip node may become decoupled from the rest of the problem in the aggregation procedure.

The rescaling is done by looking at the diagonal value of the entry before the lifting (when it's
overwritten by the value of 1.0) and rescaling with this after the call to MatZeroRowsColumns.

Currently the rescaling is always applied. Some logic may be added if we think this is expensive.