Remove csr2petsc_CreateSeqBAIJ again.
This is no longer needed as csr2petsc_CreateSeqAIJ is generalised to
setup a matrix with blocks if the group size has been set in the petsc
numbgering. It doesn't create a BAIJ matrix though - need to think about
BAIJ vs. AIJ with block size set.
It wasn't actually called from anywhere anymore and not sure it was
right either tbh.