Seperate out new create_subdomain_mesh from derive_external_subdomain_mesh.
create_subdomain_mesh only does the creation of a subdomain mesh on part
of the domain specified by an element list.
Derive_external_subdomain_mesh also does the selection of these elements
(based on prescribed_region_id options) and the mapping and inserting of
a new coordinate field into the new substate.

This is so we can hopefully reuse the routine for other jobs.

Also move generate_substate_halos from halos_registration to
halos_derivation.