Relax metric in locked (halo) regions.
In the locked regions (halo regions in parallel) and the region
immediately adjacent to it, mba can't satisfy what we ask for in the
metric This tends to upset mba and sometimes leads it to give up
altogether (leaving other regions unadapted). Therefore we adjust the
metric in the nodes of the locked regions to adhere to the locked
elements (i.e. tell mba that the mesh is perfect there already).
Directly adjacent to the locked region it will interpolate linearly
between the overwritten metric and the metric we want, so that it still
adapts to our desired quality everywhere it is allowed to.