Add a test for spherical analytical mapping.
This tests that spherical analytical mapping improves the integration of the area/volume and length/area of circular/spherical
bodies.  It also compares these results to those using both superparametric_mapping and a combination of an analytical_mapping and
superparametric_mapping (with an additional base linear case).  The combination of superparametric and analytical wins as expected.

To get this test to run several changes had to be made to the code:
 - remove the assumption that all meshes were 3d when using spherical_earth - i.e. spherical_earth is now interpretted as an
   n-sphere and circles are possible
 - remove an erroneous options_check on spherical_earth that required you to have remove pressure stabilisation even when you didn't
   have any pressure in the simulation to stabilise
 - make it possible to have both analytical_mapping and superparametric_mapping on at the same time
 - make it possible to set analytical_mapping on meshes that are external