An initial attempt at fixing shape functions and their derivatives evaluated at the facets of elements.
This work aims to replace the need for augmented shapes and is necessary with analytic sphere because the coordinates can no longer
be considered linear (similarly with superparametric).  The most obvious place this manifests itself is in the evaluation of surface
integrals of type gradient_normal where linear coordinates were required (analytical sphere would have silently returned the wrong
answer through this code path).

In the process of fixing this it became obvious that face node numbering was inconsistent between the element shape functions and
those of the facets.  To enable these to work together the element shape functions need to be reordered before being evaluated on
the surface.  Ongoing work involves further simplifying the n_s and dn_s arrays on shape functions as this reordering means that it
is no longer necessary to store a copy for every facet of an element.  However, this will first require the removal of all
augmented_shapes (so far only those in gradient_normal surface integrals have been removed) and use of dn_s and n_s outside of
reordered routines.

In particular a transform_facet_to_physical subroutine has been added that will handle everything (the reordering and evaluation of
derivatives, Jacobians etc.) in the standard way so in standard usage most of this will become hidden (hopefully).

The spherical_analytical_mapping test has been split up into a 2d and a 3d version and extra tests have been added to test the new
gradient_normal surface integral functionality.

The rest of these changes are untested so far.