Prevent isolated elements when regrowing l2 halo.
We do this by skipping elements that consist of new l2 nodes only - this is because such elements can potentially become isolated (i.e. if all adjacent elements have a non-shared node that is in an even higher level halo) Elements consisting of l2-halo nodes only, are hopefully not needed because even elements adjacent to elements adjacent to such elements will still consist of nodes that are either l1 or l2, and is therefore not owned. Thus an l2-halo node only element is never in the level 2 *element* halo.