Tools: Changes to triangle2gmsh to account for periodic boundary IDs
Identifying periodic boundaries in the triangle mesh file and
modifying the mesh object such that the gmsh writer outputs
four tags/ids for each surface element/facet. Four tags/ids are
needed by the gmsh reader in Fluidity, whereas the physical id
is the first tag, and the element owner is the fourth tag.
Otherwise the element owner from the triangle mesh is stored as
the second tag, which by default is the geometrical id in a gmsh
mesh.