BUGFIX: Multiple entries with the same element name in the schema cannot be maintained unless they have an attribute name.
The generic_scalar_field_dependency option for linear equations of state did not have a name attribute but allowed an arbitrary
number of entries to be activated.  As this already had a scalar_field_name attribute I just changed that to be the name attribute
and updated relevant tests accordingly.