Bug-fix for pseudo supermesh unittest.
When calling add_faces(), the mesh%shape gets changed - so if the
mesh is already stored in a state, we have to swap out the mesh in
state.

This also brought to my attention a refcount leak in the pseudo
supermesh code (whatever that is). Removed a spurious allocate. The
unittest now tests that all references have gone using a new function
count_references() - which also could be useful in other tests.