Fix missing flags by ensuring fldecomp build uses CPPFLAGS.
To keep things consistent with what we're doing elsewhere,
any preprocessor flags like -I and -D that are common to
the C and CXX compilers should be put in CPPFLAGS not in
CFLAGS and CXXFLAGS. In particular we dont' want to stick
PETSc's include flags in both CPPFLAGS and CXXFLAGS to avoid
duplication. That same logic should possibly also apply to
 FCFLAGS although there is some complication there by the
fact that -I is used to search both header files and
fortran module files.

This fixes #49 and closes #51