Refuse non-P1 remapping in libsam, and prevent unneccessary remapping for initial fields.
Previously DG and P2 fields where remapped to a P1 mesh before going
into sam leading to obvious degradation of the results. The only use
case where this made sense is where sam is used in flredecomp to
redecompose initial fields (which are typically not actually
discontinuous). However in that case these fields can simply be
re-prescribed from the initial value expression in the .flml, just like
zoltan does. Hook up the same logic to avoid these fields going into
sam. Flredecomping of checkpoint and any runs with parallel adaptivity
for setups with non P1 fields will not work.

Adding zoltan tags to five tests which fail with libsam but pass with zoltan.
Remove nozoltan tests:
* lagrangian_detectors had the nozoltan tag - but is no failing with sam
because it has a DG mesh and uses parallel adaptivity. So changed it
from nozoltan to zoltan - as it runs fine with zoltan.
* test_adapt_at-first-time-step had the nozoltan tag for unclear reasons.
It is a rather weak test though, and adaptivity at the first timestep is
tested in lots of other tests, so I've removed the test entirely.

Small fixes:
* Reduce a double instance of 'use detector_tools' down to one. This makes local builds on intel a lot easier.