Adding zoltan tags and some exception checking for the opions with libsam.
Adding zoltan tags to tests which fail with libsam but pass with zoltan.

Remove nozoltan tests:

lagrangian_detectors had the nozoltan tag - but is now failing with sam
because it has a DG mesh and uses parallel adaptivity. So changed it
from nozoltan to zoltan - as it runs fine with zoltan.

test_adapt_at-first-time-step had the nozoltan tag for unclear reasons.
It is a rather weak test though, and adaptivity at the first timestep is
tested in lots of other tests, so I've removed the test entirely.