Make configure fail if it can't find metis, parmetis or zoltan.
Rather than automagically switching back to sam, configure now fails if
it can't find or link with zoltan. It now also fails if failing to link
with metis or parmetis.

Also, reinstate configure test that checks we can find the zoltan.mod
fortran module and it is usable.

Remove the zoltan stuff in aclocal.m4 that was no longer in use.