Move setting of nullspaces to before KSPCreate.
This is because in petsc release it only checks upon creation of the ksp
whether there is a nullspace and then attaches it to the ksp - meaning that
if you set it later on the matrix it will ignore it. With petsc master
nullspaces are no longer attached to the ksp, so it just picks it up from the
matrix.

Also the nullspace should be set on the pmat, not mat for it to do the right
thing in the ksp (both in petsc release and master). This only matters for
the case where mat/=pmat.