Attempt at fixing new nullspace situation in petsc 3.6.
We now set the nullspace on the *mat* not the pmat. For petsc<3.6 where it picks up
the ksp nullspace from pmat, we fix this by overriding the nullspace with
KSPSetNullSpace(). For these versions, if the pmat has a nullspace, but the mat
doesn't we give an error. We do allow it with petsc 3.6 - but I don't think it's
useful in either case.

A (possibly different) nullspace is attached to *pmat* /only/ if nullspace options
are set under preconditioner::ksp. This is now only allowed if pmat/=mat,
i.e. if mat==pmat we assume the nullspace options at the level above, and
nullspace under preconditioner::ksp are verboten.