Fix weak bcs for momentum subcycling.
Previously the necessary terms associated with a weak Dirichlet bc
for advection on the rhs (the ones containing the actual bc values)
where left out entirely when using subcycling. These are now applied as
a rhs term during the subcycling.

This potentially changes any run with DG subcycling for momentum with a weak
Dirichlet inflow condition. Since the condition for the normal component
is also imposed via the continuity equation, it may have done roughly the
right thing for cases where advection is not dominant at the open
boundary. However local artefacts would still have been likely - apparent
through an increased pressure along the open boundary, and a noisy velocity.

Switch on subcycling in friction-slope-swe test. It did not pass with subcycling before this fix.