Two fixes to fix fluidity build with gcc5.
1) Do not call Zoltan_LB_Free_Part() on zero-length arrays returned by Zoltan_LB_Partition(). Not sure whether this is a bug with zoltan+gcc5 (and thus this
fix might create a small mem-leak by not deallocating zero-length
arrays). Or whether we were not supposed to call Zoltan_LB_Free_Part()
in this case in the first place.

2) in transform-to-physical, the facet_cache and full_facet_cache appear
not to do the right thing for embedded meshes: mesh_dim(X)<X%dim.
Therefore switching the cache off entirely for this case (mind you I
don't suspect this is likely to work correctly at the moment in general
since the tests are passing with gcc 4.9)