Ensure final zoltan_drive always performs a halo update.
Since we no longer interpolate in the halo regions (as we no longer lock
the l2 halo, the new halo is not guaranteed to fall within the old
halo), a halo update is necessary after the interpolation. Normally this
happens in zoltan when the newly redistributed state gets populated.
However in some cases zoltan decides no changes in the distribution is
necessary, in this case we need to call halo update by hand to ensure
the halos are consistent before going into the subsequent timestep.