Removing the radial extrusion code and reimplementing it in the previously just planar code.
The radial extrusion code was mostly a copied and pasted version of the original "planar" code with only minor modifications.  This
was unnecessary and made implementing new extrusion functionality complicated.  This commit streamlines this a bit by merging the
two code paths.

This has the bonus of adding some missing functionality to spherical, including surface heights for maps and proper sigma layering (though this is not
implemented in the most efficient manner).