Change exit state following successful test runs
Prior to this commit, testharness can run successfully but still exit with error if tests it runs return failures. This can be confusing as it suggests that testharness itself has experienced an error - a state which occurs if (for example) an invalid xml file is specified in the input. This commit removes the errant behaviour, and puts the onus on output parsers to detect whether or not tests are passing.