One last nitpick: clarify construct_les_dg() interface.
If a reference counted object is passed as an intent(inout) argument,
I expect its value to be changed - not the IN reference to be silently
dropped and replaced by a new one. Subroutines should make sense on
their own, wherever they are called from, so if the IN argument is something
different than the OUT argument, they should just be separate arguments.

* remove unused variables