Change purely in test options:
Adding 'vlong' as a valid test length. The intention is that this is used for
tests which are not suitable for running on a workstation or single cluster
node, defined in this case as tests which run without output to stdout for more
than three hours on a system with 8 x 2.4GHz hyperthreaded Xeon cores and 24GB
RAM.

The length is chosen to allow later addition of 'make longtest' which is
intended to complete on the above system spec in less than one week, with áll
'long' tests run sequentially. At a practical level, this allows testing of
'long' tests on a weekly basis within local CI infrastructure, 'vlong' tests
being run remotely on larger compute capability.