Fix extrude with general (not 2plus1) adaptiivty in parallel.
Zoltan needed to be told which mesh to redistribute (and derive the
other meshes from). Understandably it was thinking this is the external
mesh, however in the case of an extruded mesh with "general" adaptivity
we actually derive everything from the 3d adapted mesh, and leave the external
horizontal mesh untouched (i.e. it keeps the old decomposition). We
should probably actually remove the horizontal mesh, and make the 3d
adapted mesh the official external mesh in the options tree.