Reviving parallel periodic adaptivity and adding a test.
This required various fixes:
 - create_subdomain_mesh was wrong for the case with discontinuous facets
 - zoltan needed to catch periodic surface ids before constructing the mesh (rather than just setting them to 0 after)
 - numerous places (mostly associated with strip l2 halo) needed periodic or multivalued halo flags setting

The test is just a parallel version of a preexisting periodic adaptivity case.  It runs on 3 processes and tries to use a high
aspect ratio domain to avoid wrap around failures.  However if it does hit a wrap around failure all tests pass and instead a
warning is issued.