Adding calculated arrays of states to keep track of previously evaluated dependencies in momentum diagnostic fields.
Also providing routines to update the two arrays (of all states and of states in the current phase).  These need to be called
appropriately after every diagnostic variable call that gets changed to also calculate dependencies.  The current use is pointless
(because the field after which it is called is not actually calculating any dependencies) and only intended as an example.