Switch to calculating the dependencies of diagnostic variables in momentum diagnostic fields.
This tracks the dependencies through a calculated mask, which requires some copying back and forward between two arrays of states (one for all states the other for the phase).  This commit includes a bug fix for this to ensure that the copied fields are incref'd.

While fixing this a bug was also noticed in some similar code in diagnostics where calculated dependencies were being aliased to all states (potentially leading to errors with multi-blah simulations).  Implementing a suggested fix for that here.