Making petsc_readnsolve special
At the moment all the Redhat-based test queues on buildbot, as well as the xenial queue, are being intermittently stalled by petsc_readnsolve hanging. This change sets the test to 'special', removing it from the default queue. Buildbot has been configured to add a petsc_readnsolve test to the trusty queue, run explicitly at the end of the testing, so petsc_readnsolve remains tested on the currently core platform as of 2015/03/30.