ALTurbine Airfoils
The alturbine_airfoils module has been added
you may initialise the airfoil table data from
the alturbine. For each turbine the available
airfoils will be added and stored. The blade
thickness to chord, Reynolds number and angle of
attack will be used to interpolate (using linear interpolation)
in order to compute the correct Lift and Drag coefficients. This is
before the dynamic stall model is used.