Turbine Operation routines
1) Turbine rotates around a prespecified axis of rotation and
normal vector to the plane of rotation. This is the rotate_turbines
subroutine which is part of the ALTurbine module and takes as an
input only the angle of rotation theta in rad.
2) At each element of the blade a Force coefficient term is assigned
which will be used in order to compute the momentum sink terms.
3) The force coefficients will be computed in terms of CFn, CFt and CFs
so they need to be transformed to the mesh system of coordinates x,y,z
as CFx, CFy and CFz before the forcing terms are applied.