Fix race condition in writing pvtus.
The hack that allows us to write the vtus of a pvtu into a
sub-directory, involves the following step:
1) create the subdirectory, and tell vtk to write the pvtu and vtus into
  that subdirectory
2) move the pvtu to the directory above
3) fix the paths in the pvtus to relative paths into the subdirectory

Steps 2) and 3) are done by process 0. But currently no steps are taken
to ensure the other processes are actually finished with 1).

This change puts a barrier between 2) and 3) to avoid a race condition
that causes the pvtu to get mangled (observed on Debian unstable with
vtk 6.3).