Fix vector fields in linear_interpolation_state:
Previously assumed the ordering of vector fields in old_state and new_state is the same, and
identified coordinate fields by looking at the name of the vector fields instead of the name under
which it was inserted. The latter breaks linear_interpolation_scalar/vector/tensor:

call linear_interpolat(sfield_old, old_positions, sfield_new, new_positions)

which creates a temporary old_state where old_positions is inserted as "Coordinate". This should still
work if old_positions%name is not "Coordinate" or "MeshNameCoordinate".