Fixing cache_solver_context solver option.
Not only was it not doing anything, with the ksp
just being overwritten every solver call; It therefore also
leaked memory all over the place. Fixing this makes
the balance-pressure test run twice as fast, probably
mostly due to CPU memory cache misalignment as a consequence
of the memory leak.