When reusing cached ksp still call setup_ksp_from_options.
Since we don't call KSPSetOperators() again, PETSc should be smart
enough not unnecessarily redo setup work. This ensures that if options
are set differently in subsequent calls to petsc_solve() they aren't
silently ignored. For instance if the inner (velocity) solve of a
FullMomentumMatrix or FullMassMatrix schur solve is reused in the
 correction solve - PETSc sets non-zero-initial-guess to false in
the inner schur solve, so it needs to be reset in the correction solve.

Rename SetupKSP to create_ksp_from_options() to better reflect what it
does: call KSPCreate(), KSPSetOperators() and then
setup_ksp_from_options().