Make bc smoothing work in parallel
by adding halos to the surface mesh the values are stored on. This
reuses generate_free_surface_halos(), now renamed as
generate_free_surface_mesh_halos() and moved to halos_derivation.

This commit has many unrelated updates to Makefile.dependencies, as the
Makefile.dependencies on master are currently out of sync!