Fix bugs related to intersect_elements() short-cut.
The optimisation in commit c68550e5d8fb700ecf1c6acda28f66b00eb2d465
introduced a short-cut in intersect_elements() where if the intersection
was found to be empty it no longer bothered to allocate a valid
0-element mesh. This condition had to then be catched
by always checking with has_references() on the returned mesh. On second
thought this is a rather flaky solution. If a function returns an
allocated femtools object it should always return a valid object - or
there should be some other mechanism to indicate the function result
should be discarded (e.g. extract_scalar_field with a stat argument).

In this spirit, intersect_elements() now has an optional
empty_intersection argument. The short-cut is only invoked if this
argument is provided  - thus restoring the semantics from before the
introduction of the short-cut if the argument is not present,
which fixes a few places where intersect_element() calls where not
checked with has_references(). If the argument is provided,
upon return empty_intersection can be checked whether the function
result is valid or not.

This fixes a.o. unifiedmesh and therefore the
explicit-hyperc-shear-adapt test. The short-cut is now also extended to
the libsupermesh implementation (although I believe this is currently
entirely untested?).